Features APPSTUD-01, APPSTUD-02, APPSTUD-03
Following features have been added to this commit.
1. The current user location will be displayed on the map if available.

2. The places will be displayed around you as pins on the map. Each pin
is a circle with a black border and the picture of the place at the
center. The picture will be retrieved thanks to an API request (more
information will be provided about that later in this document).

A button will allow the user to center the map on his location if
available (it’s not included in the screenshot)